PARQUET-2161: Fix row index generation
The row indexes introduced in PARQUET-2117 are not computed correctly
when:
(1) range or offset metadata filter is applied, and
(2) the first row group was eliminated by the filter

For example, if a file has two row groups with 10 rows each, and we
attempt to only read the 2nd row group, we are going to produce row
indexes 0, 1, 2, ..., 9 instead of expected 10, 11, ..., 19.

This happens because functions `filterFileMetaDataByStart`
and `filterFileMetaDataByMidpoint` modify their input `FileMetaData`.
To return correct result, `generateRowGroupOffsets` has to be computed
before these filters are applied.