PARQUET-2260: Bloom filter size shouldn't exceed the configured maxBytes (#1043)
If parquet.bloom.filter.max.bytes configuration is not a power of 2 value, the size of the bloom filter generated will exceed this value. For example, if set parquet.bloom.filter.max.bytes as 1024 * 1024+1= 1048577 , the bytes size of bloom filter generated will be 1024 * 1024 * 2 = 2097152. This does not match the definition of the parameter. This commit fixes the issue.