PARQUET-2429: Reduce direct input buffer churn
Currently input buffers are grown one chunk at a time as the
compressor or decompressor receives successive setInput calls. When
decompressing a 64MB block using a 4KB chunk size, this leads to
thousands of allocations and deallocations totaling GBs of memory.
By growing the buffer 2x each time, we avoid this and instead use a
modest number of allocations.