PARQUET-2171. Support vectored IO in Hadoop stream reads
Bridge takes a ByteBufferAllocator as argument, rather than an allocate method.
This lines up for future plans of mine to support passing down of
a free() call to hadoop fs api; this will let us return a buffer on failures,
especially off-head stuff, rather than wait for GC.

Also taken as a parameter to decide whether or not the API
should be used, based on the isDirect() flag.

Test cleanup to go with.

Change-Id: Ib90bc9ee49f5db7f12eff1c73b599b18fba60ecf