GH-3083: Make DELTA_LENGTH_BYTE_ARRAY default encoding for binary
The current default for V1 pages is PLAIN encoding. This encoding mixes string length with string data.
This is inefficient for for skipping N values, as the encoding does not allow random access.
It's also slow to decode as the interleaving of lengths with data does not allow efficient batched
implementations and forces most implementations to make copies of the data to fit the
usual representation of separate offsets and data for strings.

DELTA_LENGTH_BYTE_ARRAY has none of the above problems as it separates offsets and data.
The parquet-format spec also seems to recommend this
https://github.com/apache/parquet-format/blob/c70281359087dfaee8bd43bed9748675f4aabe11/Encodings.md?plain=1#L299